some tests updated, some refatored
