+++ tests/Endpoints/CountriesTest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace UON\Tests\Endpoint;
+namespace UON\Tests\Endpoints;
 
 use PHPUnit\Framework\TestCase;
 use UON\Config;
-use UON\Endpoint\Cities;
+use UON\Endpoints\Countries;
 
-class CitiesTest extends TestCase
+class CountriesTest extends TestCase
 {
-    private $_cities;
-    private $_city;
+    private $_countries;
+    private $_country;
 
-    public static $citiesId;
+    public static $countryId;
 
     public function __construct($name = null, array $data = [], $dataName = '')
     {
@@ -19,31 +19,30 @@ class CitiesTest extends TestCase
         $config = new Config();
         $config->set('token', file_get_contents(__DIR__ . '/../_token.txt'));
 
-        $this->_cities = new Cities($config);
-        $this->_city = array(
-            'country_id' => '1',
+        $this->_countries = new Countries($config);
+        $this->_country = [
             'name' => 'Кингконгстоунт',
             'name_en' => 'Kinkongstoun'
-        );
+        ];
     }
 
     public function testCreate()
     {
-        $result = $this->_cities->create($this->_city);
-        self::$citiesId = $result['message']->id;
+        $result = $this->_countries->create($this->_country);
+        self::$countryId = $result['message']->id;
         $this->assertInternalType('array', $result);
     }
 
     public function testRead()
     {
-        $result = $this->_cities->all($this->_city['country_id']);
+        $result = $this->_countries->all();
         $this->assertInternalType('array', $result);
     }
 
     public function testUpdate()
     {
-        $update = $this->_cities->update(self::$citiesId, $this->_city);
-        $this->assertInternalType('array', $update);
+        $result = $this->_countries->update(self::$countryId, $this->_country);
+        $this->assertInternalType('array', $result);
     }
 
 }
