+++ tests/Endpoints/HotelsTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace UON\Tests\Endpoint;
+namespace UON\Tests\Endpoints;
 
 use PHPUnit\Framework\TestCase;
 use UON\Config;
-use UON\Endpoint\Hotels;
+use UON\Endpoints\Hotels;
 
 class HotelsTest extends TestCase
 {
