+++ tests/Endpoints/LeadsTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace UON\Tests\Endpoint;
+namespace UON\Tests\Endpoints;
 
 use PHPUnit\Framework\TestCase;
 use UON\Config;
-use UON\Endpoint\Leads;
+use UON\Endpoints\Leads;
 
 class LeadsTest extends TestCase
 {
