+++ tests/Endpoints/SourcesTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace UON\Tests\Endpoint;
+namespace UON\Tests\Endpoints;
 
 use PHPUnit\Framework\TestCase;
 use UON\Config;
-use UON\Endpoint\Sources;
+use UON\Endpoints\Sources;
 
 class SourcesTest extends TestCase
 {
